+++ webpack.config.js
@@ -7,52 +7,55 @@ const CopyPlugin = require("copy-webpack-plugin");
 const ZipPlugin = require("zip-webpack-plugin");
 
 module.exports = (env, argv) => {
-    const isPro = argv.mode === "production";
+    const production = argv.mode === "production";
     const plugins = [
         new MiniCssExtractPlugin({
-            filename: isPro ? "dist/index.css" : "index.css",
-        })
+            filename: production ? "dist/index.css" : "index.css",
+        }),
     ];
-    let entry = {
-        "index": "./src/index.ts",
-    };
-    if (isPro) {
-        entry = {
-            "dist/index": "./src/index.ts",
-        };
-        plugins.push(new webpack.BannerPlugin({
-            banner: () => {
-                return fs.readFileSync("LICENSE").toString();
-            },
-        }));
-        plugins.push(new CopyPlugin({
-            patterns: [
-                {from: "preview.png", to: "./dist/"},
-                {from: "icon.png", to: "./dist/"},
-                {from: "README*.md", to: "./dist/"},
-                {from: "plugin.json", to: "./dist/"},
-                {from: "src/i18n/", to: "./dist/i18n/"},
-            ],
-        }));
-        plugins.push(new ZipPlugin({
-            filename: "package.zip",
-            algorithm: "gzip",
-            include: [/dist/],
-            pathMapper: (assetPath) => {
-                return assetPath.replace("dist/", "");
-            },
-        }));
+    if (production) {
+        plugins.push(
+            new webpack.BannerPlugin({
+                banner: () => {
+                    return fs.readFileSync("LICENSE").toString();
+                },
+            }),
+        );
+        plugins.push(
+            new CopyPlugin({
+                patterns: [
+                    {from: "preview.png", to: "./dist/"},
+                    {from: "icon.png", to: "./dist/"},
+                    {from: "README*.md", to: "./dist/"},
+                    {from: "plugin.json", to: "./dist/"},
+                    {from: "src/i18n/", to: "./dist/i18n/"},
+                    {from: "dist/kernel.js", to: "./dist/"},
+                ],
+            }),
+        );
+        plugins.push(
+            new ZipPlugin({
+                filename: "package.zip",
+                algorithm: "gzip",
+                include: [/dist/],
+                pathMapper: (assetPath) => {
+                    return assetPath.replace("dist/", "");
+                },
+            }),
+        );
     } else {
-        plugins.push(new CopyPlugin({
-            patterns: [
-                {from: "src/i18n/", to: "./i18n/"},
-            ],
-        }));
+        plugins.push(
+            new CopyPlugin({
+                patterns: [
+                    {from: "src/i18n/", to: "./i18n/"},
+                ],
+            }),
+        );
     }
     return {
         mode: argv.mode || "development",
-        watch: !isPro,
-        devtool: isPro ? false : "eval",
+        watch: !production,
+        devtool: production ? false : "eval-source-map",
         output: {
             filename: "[name].js",
             path: path.resolve(__dirname),
@@ -64,9 +67,11 @@ module.exports = (env, argv) => {
         externals: {
             siyuan: "siyuan",
         },
-        entry,
+        entry: {
+            [production ? "dist/index" : "index"]: "./src/index.ts",
+        },
         optimization: {
-            minimize: true,
+            minimize: production,
             minimizer: [
                 new EsbuildPlugin(),
             ],
@@ -84,7 +89,7 @@ module.exports = (env, argv) => {
                             loader: "esbuild-loader",
                             options: {
                                 target: "es6",
-                            }
+                            },
                         },
                     ],
                 },
@@ -100,7 +105,7 @@ module.exports = (env, argv) => {
                             loader: "sass-loader", // compiles Sass to CSS
                         },
                     ],
-                }
+                },
             ],
         },
         plugins,